Factor out signed out reminder
